ref(JitsiMeetLogStorage): move to base/logging
