+++ modules/UI/shared_video/SharedVideo.js
@@ -301,8 +301,7 @@ export default class SharedVideoManager {
 
                 // FIXME The cat is out of the bag already or rather _room is
                 // not private because it is used in multiple other places
-                // already such as AbstractPageReloadOverlay and
-                // JitsiMeetLogStorage.
+                // already such as AbstractPageReloadOverlay.
                 conference: APP.conference._room,
                 id: self.url,
                 isFakeParticipant: true,
